% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_klamath_rivers_line}
\alias{all_klamath_rivers_line}
\title{All Klamath River Lines}
\format{
A sf df with 20 rows and 2 columns
}
\usage{
all_klamath_rivers_line
}
\description{
National Hydrography Dataset of flowlines, processed and cleaned. Includes: Blue Creek, Bogus Creek, Camp Creek, Clear Creek, Fall Creek, Indian Creek, Jenny Creek, Klamath River, Link River, Lost River, Salmon River, Scotch Creek, Scott River, Shasta River, Shovel Creek, Spencer Creek, Sprague River, Trinity River, Williamson River, Wood River
}
\keyword{datasets}
