% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_basin.R
\name{assign_sub_basin}
\alias{assign_sub_basin}
\title{Asign Sub Basin}
\usage{
assign_sub_basin(
  data,
  sub_basin,
  is_point = TRUE,
  lon_col = "longitude",
  lat_col = "latitude",
  sub_basin_col = "NAME"
)
}
\arguments{
\item{data}{A data frame or `sf` object with location data.}

\item{sub_basin}{An `sf` polygon object containing sub-basin geometries and names.}
}
\value{
A data frame (or sf object) with a new `sub_basin` column in lowercase.
}
\description{
This function takes an `sf` object containing sub basin location in the Klamath Basin (e.g., trinity, upper klamath, etc)
and assignes a sub-basin to each data point.
}
