% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_river_miles.R
\name{get_river_miles}
\alias{get_river_miles}
\title{Get river miles}
\usage{
get_river_miles(
  river_layer,
  river_name,
  interval_m = 10,
  downstream_end = "end",
  crs_proj = 32610
)
}
\description{
Get river miles
}
