% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klamath_hucs}
\alias{klamath_hucs}
\title{Klamath River HUCs}
\format{
A sf df with 12 rows and 3 columns
\itemize{
  \item \code{huc8}: 8 digit HUC code
  \item \code{name}: HUC name
  \item \code{geometry}: sf polygon
  }
}
\usage{
klamath_hucs
}
\description{
National Hydrography Dataset of HUCs, processed and cleaned. Source: https://www.sciencebase.gov/catalog/item/5d30c29ae4b01d82ce84aa5e
}
\keyword{datasets}
