% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhd_catchments}
\alias{nhd_catchments}
\title{NHD Catchments}
\format{
A tibble with 4591 rows and 13 columns
\itemize{
  \item \code{nhdplusid}: unique identifier for the NHD dataset
  \item \code{areasqkm}: area of catchment (sq km)
  \item \code{geometry}: sf geometry
  }
}
\usage{
nhd_catchments
}
\description{
National Hydrography Dataset of catchments, filtered to nhdplusids for rivers included in the Klamath Basin Science Collaborative. Source: https://www.epa.gov/waterdata/get-nhdplus-national-hydrography-dataset-plus-data
}
\keyword{datasets}
