% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed_nhd}
\alias{processed_nhd}
\title{NHD Flowlines}
\format{
A tibble with 4236 rows and 73 columns
}
\usage{
processed_nhd
}
\description{
National Hydrography Dataset of flowlines, filtered to nhdplusids for rivers included in the Klamath Basin Science Collaborative. Source:https://www.sciencebase.gov/catalog/item/5d30c29ae4b01d82ce84aa5e
}
\keyword{datasets}
