% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_river_mile.R
\name{find_nearest_river_miles}
\alias{find_nearest_river_miles}
\title{Find Nearest River Miles}
\usage{
find_nearest_river_miles(points_sf)
}
\arguments{
\item{points_sf}{An `sf` object containing points with a `stream_short` column.}
}
\value{
An `sf` object with an added `nearest_river_mile` column.
If no river mile data exists, the value will be `NA`.
}
\description{
This function takes an `sf` object containing point locations (e.g., sampling points)
and finds the nearest river mile for each point.
}
